new observabiliyt UI update
