+++ .changeset/humble-plums-feel.md
@@ -1,0 +1,5 @@
+---
+'@mastra/playground-ui': patch
+---
+
+Moved timeline span components from logs domain to traces domain, added missing memory and workspace span type support, added span type color legend to trace timeline, and fixed scroll-into-view when navigating spans with Next/Prev
