+++ packages/playground-ui/src/domains/logs/components/trace/get-descendant-ids.ts
@@ -1,25 +1,1 @@
-import type { UISpan } from './types';
-
-export function getSpanDescendantIds(span: UISpan): string[] {
-  if (!span.spans || span.spans.length === 0) {
-    return [];
-  }
-
-  const descendantIds: string[] = [];
-
-  span.spans.forEach(childSpan => {
-    descendantIds.push(childSpan.id);
-    descendantIds.push(...getSpanDescendantIds(childSpan));
-  });
-
-  return descendantIds;
-}
-
-export function getAllSpanIds(spans: UISpan[]): string[] {
-  const ids: string[] = [];
-  for (const span of spans) {
-    ids.push(span.id);
-    ids.push(...getSpanDescendantIds(span));
-  }
-  return ids;
-}
+export { getSpanDescendantIds, getAllSpanIds } from '@/domains/traces/hooks/get-all-span-ids';
