+++ packages/playground-ui/src/domains/logs/components/trace/shared.tsx
@@ -1,55 +1,1 @@
-import { BrainIcon } from 'lucide-react';
-import type { UISpanStyle } from './types';
-import { AgentIcon, McpServerIcon, ToolsIcon, WorkflowIcon } from '@/ds/icons';
-
-export const spanTypePrefixes = ['agent', 'workflow', 'model', 'mcp', 'tool', 'other'];
-
-export function getSpanTypeUi(type: string) {
-  const typePrefix = type?.toLowerCase().split('_')[0];
-
-  const spanTypeToUiElements: Record<string, UISpanStyle> = {
-    agent: {
-      icon: <AgentIcon />,
-      color: 'oklch(0.75 0.15 250)',
-      label: 'Agent',
-      bgColor: 'oklch(0.75 0.15 250 / 0.1)',
-      typePrefix: 'agent',
-    },
-    workflow: {
-      icon: <WorkflowIcon />,
-      color: 'oklch(0.75 0.15 200)',
-      label: 'Workflow',
-      bgColor: 'oklch(0.75 0.15 200 / 0.1)',
-      typePrefix: 'workflow',
-    },
-    model: {
-      icon: <BrainIcon />,
-      color: 'oklch(0.75 0.15 320)',
-      label: 'Model',
-      bgColor: 'oklch(0.75 0.15 320 / 0.1)',
-      typePrefix: 'model',
-    },
-    mcp: {
-      icon: <McpServerIcon />,
-      color: 'oklch(0.75 0.15 160)',
-      label: 'MCP',
-      bgColor: 'oklch(0.75 0.15 160 / 0.1)',
-      typePrefix: 'mcp',
-    },
-    tool: {
-      icon: <ToolsIcon />,
-      color: 'oklch(0.75 0.15 100)',
-      label: 'Tool',
-      bgColor: 'oklch(0.75 0.15 100 / 0.1)',
-      typePrefix: 'tool',
-    },
-  };
-
-  if (typePrefix in spanTypeToUiElements) {
-    return spanTypeToUiElements[typePrefix];
-  }
-
-  return {
-    typePrefix: 'other',
-  };
-}
+export { spanTypePrefixes, getSpanTypeUi } from '@/domains/traces/components/shared';
