+++ packages/playground-ui/src/domains/traces/components/timeline-name-col.tsx
@@ -1,0 +1,54 @@
+import { TimelineStructureSign } from './timeline-structure-sign';
+import type { UISpan, UISpanStyle } from '../types';
+import { cn } from '@/lib/utils';
+
+type TimelineNameColProps = {
+  span: UISpan;
+  spanUI?: UISpanStyle | null;
+  isFaded?: boolean;
+  depth?: number;
+  onSpanClick?: (id: string) => void;
+  selectedSpanId?: string;
+  isLastChild?: boolean;
+  hasChildren?: boolean;
+  isRootSpan?: boolean;
+  isExpanded?: boolean;
+};
+
+export function TimelineNameCol({
+  span,
+  spanUI: _spanUI,
+  isFaded,
+  depth = 0,
+  onSpanClick,
+  selectedSpanId,
+  isLastChild,
+  hasChildren: _hasChildren,
+  isRootSpan,
+  isExpanded: _isExpanded,
+}: TimelineNameColProps) {
+  return (
+    <div
+      data-span-id={span.id}
+      aria-label={`View details for span ${span.name}`}
+      className={cn('rounded-md transition-colors flex opacity-80 min-h-8 items-center rounded-l-lg', {
+        'opacity-30 [&:hover]:opacity-60': isFaded,
+        'bg-surface4': selectedSpanId === span.id,
+      })}
+      style={{ paddingLeft: `${depth * 1}rem` }}
+    >
+      {!isRootSpan && <TimelineStructureSign isLastChild={isLastChild} />}
+
+      <button
+        onClick={() => onSpanClick?.(span.id)}
+        className={cn(
+          'text-ui-sm flex items-center text-left gap-1.5 text-neutral6 w-full min-w-0 rounded-md h-full px-2 py-1 transition-colors',
+          '[&>svg]:transition-all [&>svg]:shrink-0 [&>svg]:opacity-0 [&>svg]:w-[1em] [&>svg]:h-[1em] [&>svg]:ml-auto',
+          'hover:bg-surface4 [&:hover>svg]:opacity-60',
+        )}
+      >
+        <span className="min-w-0 truncate">{span.name}</span>
+      </button>
+    </div>
+  );
+}
